Rend_ApplyLightAdaptation now works with int values. Added  Rend_GetLightAdaptVal same as Rend_ApplyLightAdaptation but returns the offset instead of applying it to the variable,

+++ doomsday/Src/rend_decor.c
@@ -293,7 +293,7 @@ boolean Rend_CheckDecorationBounds(fixed_t bounds[6], float fMaxDist)
  */
 float Rend_CheckSectorLight(sector_t *sector, ded_decorlight_t * lightDef)
 {
-    byte    lightlevel;
+    int    lightlevel;
     float   factor;
 
     lightlevel = sector->lightlevel;
